refactor(app): extract action -> event conversion to action_events (phase 3i)
Move action_to_events and handle_visual_line_movement out of render.rs
into a new app/action_events.rs sibling. ~265 lines.

These bridge the Action enum (key-press semantics) to the Event stream
(buffer mutation), routing soft-wrapped movement through the cached
layout for visual-line correctness.

render.rs goes from 1894 -> 1630 lines (-264). 2026 tests still pass.
